--- README.md
@@ -1,2 +1,0 @@
-# TA-2-independence24
- official work of TA2 from independence24
